frontend various fixes after tests

+++ XYZEngineeringProject.UI/src/app/components/client/client-form-new/client-form-new.component.ts
@@ -106,7 +106,7 @@ export class ClientFormNewComponent implements OnInit {
 
   addClientContact() {
     const group = this.fb.group({
-      firstName: [''],
+      firstName: ['', Validators.required],
       surname: [''],
       phone: [''],
       email: [''],
@@ -118,15 +118,14 @@ export class ClientFormNewComponent implements OnInit {
 
   removeClientContact(index: number) {
     if(confirm(this.translateService.instant('Alert.deleteContactPerson') + this.clientContacts.value[index].firstName + " " + this.clientContacts.value[index].surname + "?")) {
-      this.clientService.deleteClientContact(this.clientDetails, index).subscribe({
+      if (this.editMode == true && this.clientDetails.contacts) {
+        this.clientService.deleteClientContact(this.clientDetails, index).subscribe({
         next: (res) => {
-          this.clientContacts.removeAt(index);
-        },
-        error: (error) => {
-          let msg = 'Error'
-          window.alert(msg);
         }
-      })      
+      })     
+    } 
+      this.clientContacts.removeAt(index);
+
     } 
   }
